% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCalendarHeatmap.R
\name{.GetBoundaries}
\alias{.GetBoundaries}
\title{GetBoundaries}
\usage{
.GetBoundaries(dates, first = FALSE)
}
\arguments{
\item{dates}{Date, the dates that belong to the interval. Can also pass just
the first and last date.}

\item{first}{logical, is this the first of a set of intervals? When TRUE, the
function will return a path that encloses the dates. When \code{FALSE}, the function
will return a path that encloses the dates on 3 sides only. This is useful
because it prevents overplotting when one interval's right boundary is
another interval's left boundary.}
}
\value{
data.frame, containing x and y values to be used in a call to the
\code{ggplot::geom_path} function
}
\description{
This function takes a set of dates and returns the set of coordinates that
can be used to draw a polygon around the specified set of dates when arranged
in a grid as in the \code{CalendarHeatmap} function.
}
