% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCalendarHeatmap.R
\name{.GetPath}
\alias{.GetPath}
\title{Get Path}
\usage{
.GetPath(plot.data)
}
\arguments{
\item{plot.data}{data.table, containing the data that will be used to
construct the calendar plot}
}
\value{
data.table containing the path between the different months
}
\description{
Finds the x and y values needed to construct a path to separate adjacent
months in the plot.
}
