% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCalendarHeatmap.R
\name{.MinMaxNorm}
\alias{.MinMaxNorm}
\title{MinMaxNorm}
\usage{
.MinMaxNorm(vector)
}
\arguments{
\item{vector}{numeric, vector of values to be scaled}
}
\value{
numeric vector equal in length to the vector provided, with all
values scaled to be between 0 and 1 (inclusive).
}
\description{
Internal function for scaling user provided size values to be between 0 and 1 (inclusive).
}
