% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCalendarHeatmap.R
\name{MakePlot}
\alias{MakePlot}
\title{Make Plot}
\usage{
MakePlot(data, date.column, fill, bg.col, size, missing.is.zero = TRUE,
  day.border.col, month.border.col)
}
\arguments{
\item{data}{data.frame, containing the data to be used to create the plot}

\item{date.column}{character, the name of the column in `data` that contains the dates}

\item{fill}{character, the name of the column in `data` that will be used to determing the colors in the heatmap}

\item{bg.col}{character, the background color for the plot}

\item{size}{character or numeric, either the column to be used to determine the size of the squares, or a number that defines the size to use for all squares. Optional, and at present not recommended.}

\item{missing.is.zero}{logical, should dates for which no data is presented be assumed to be zero?}

\item{day.border.col}{character, the color to be used for the boundaries between days}

\item{month.border.col}{character, the color to be used for the boundaries between months}
}
\value{
ggplot object containing a calendar heatmap
}
\description{
Make Plot
}
