% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PlotCalendarHeatmap.R
\name{PlotCalendarHeatmap}
\alias{PlotCalendarHeatmap}
\title{PlotCalendarHeatmap}
\usage{
PlotCalendarHeatmap(data, bg.col)
}
\arguments{
\item{data}{data.frame, which must have a column named \code{date} containing
dates, and at least one other column.}
}
\description{
PlotCalendarHeatmap
}
